requirements.txt
click
python-dotenv
ooo-dev-tools
numpy
tqdm
